feat(tests): add user lifecycle integration tests
- Introduced a comprehensive suite of integration tests for user lifecycle, including user creation, login, logout, profile updates, and deletion.
- Utilized unique user data generation to avoid conflicts during test execution.
- Ensured proper server setup and teardown to maintain test isolation and reliability.
